create DTO Request and Response
