+++ src/main/java/com/rental/PropertyRentalApi/Controller/AuthController.java
@@ -6,11 +6,6 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.rental.PropertyRentalApi.DTO.AuthRequest;
-import com.rental.PropertyRentalApi.DTO.AuthResponse;
-import com.rental.PropertyRentalApi.Service.AuthService;
-
-
 
 @RestController
 @RequestMapping("/auth")
