--- src/main/java/com/rental/PropertyRentalApi/DTO/AuthResponse.java
@@ -1,18 +1,0 @@
-package com.rental.PropertyRentalApi.DTO;
-
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class AuthResponse {
-    private String username;
-    private String fullname;
-    private String email;
-    private String token; 
-    
-}
-
