+++ src/main/java/com/rental/PropertyRentalApi/DTO/request/AuthRequest.java
@@ -1,0 +1,10 @@
+package com.rental.PropertyRentalApi.DTO.request;
+
+import lombok.Data;
+
+@Data
+public class AuthRequest {
+        private String email;
+        private String password;
+    }
+
