+++ src/main/java/com/rental/PropertyRentalApi/DTO/request/RegisterRequest.java
@@ -1,0 +1,11 @@
+package com.rental.PropertyRentalApi.DTO.request;
+
+import lombok.Data;
+
+@Data
+public class RegisterRequest {
+    private String fullname;
+    private String username;
+    private String email;
+    private String password;
+}
