+++ src/main/java/com/rental/PropertyRentalApi/DTO/response/RegisterResponse.java
@@ -1,0 +1,13 @@
+package com.rental.PropertyRentalApi.DTO.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class RegisterResponse {
+    private String message;
+    private UserResponse user;
+}
