+++ src/main/java/com/rental/PropertyRentalApi/DTO/response/UserResponse.java
@@ -1,0 +1,18 @@
+package com.rental.PropertyRentalApi.DTO.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class UserResponse {
+
+    // We only return safe fields (NOT password)
+    private Long id;
+    private String fullname;
+    private String username;
+    private String email;
+    private String role;
+}
