+++ src/main/java/com/rental/PropertyRentalApi/Repository/UserRepository.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<UserEntity, Long> {
-	 boolean existsByUsername(String username);
+        Optional<UserEntity> findByEmail(String email);
 	    Optional<UserEntity> findByUsername(String username);
 
 }
