+++ src/main/java/com/rental/PropertyRentalApi/Service/Impl/AuthServiceImpl.java
@@ -1,9 +1,12 @@
-package com.rental.PropertyRentalApi.Service;
+package com.rental.PropertyRentalApi.Service.Impl;
 
 import com.rental.PropertyRentalApi.DTO.request.RegisterRequest;
 import com.rental.PropertyRentalApi.DTO.response.RegisterResponse;
+
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
-public interface AuthService {
-    RegisterResponse register(RegisterRequest request);
+@Service
+@RequiredArgsConstructor
+public class AuthServiceImpl {
 }
